requirements.txt
google-adk>=1.0.0
httpx>=0.24.0
python-dotenv>=1.0.0
